+++ config.ts
@@ -1,4 +1,6 @@
-export const API_URL = `api.merrymake.io`;
+export const MERRYMAKE_IO = `merrymake.io`;
+export const API_URL = `api.${MERRYMAKE_IO}`;
+export const RAPIDS_HOST = `https://rapids.${MERRYMAKE_IO}`;
 
 export const HTTP_HOST = `https://${API_URL}`;
 export const SSH_HOST = `${API_URL}`;